get wireguard working for ipv4, disable ipv6
After some tcpdumping at various hops I got wireguad to suddenly work
afer dis- and reenabeling some settings and rebooting the router.

The issue before this was that when I pinged devices inside the LAN
from mutalisk connected to nydus via wireguad, the nydus would not
route the reply to the ping from its physical interface to wg0.

It does this now for IPv4. For IPv6 the destination hosts still do not
even send a reply. I think here I have a more general issue with how I
need to configure the LAN so that devices know they can reach another
local subnet via nydus.

So I disabled IPv6 in the wireguard tunnel for now, so that connections
do not try to use it and get stuck. That way I have a fully functional config for now.