make nix build pure
All configurations build without the --impure flag now.
The only file access at build time has been replaced with a special
flake input under a local path.
This is pure because of the hash that is attached to the input and
secure because of a salt, some entropy, that is added to the input.